+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * raised minimum required PHP version to 5.6.0
   * added support for argument verification with bovigo/assert
   * added support for PHP 7 return type hints
+  * `bovigo\callmap\Verification` is now automatically blacklisted in PHPUnit and will not appear in PHPUnit error stacks any more
 
 
 1.1.0 (2015-10-05)